+++ lib/command/src/Obelisk/Command/Project.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TemplateHaskell #-}
 module Obelisk.Command.Project
   ( InitSource (..)
   , initProject
@@ -23,12 +24,14 @@ import Data.Bits
 import Data.Function (on)
 import qualified Data.Text as T
 import System.Directory
+import System.Environment (lookupEnv)
 import System.FilePath
 import System.IO.Temp
 import System.IO.Unsafe (unsafePerformIO)
 import System.Posix (FileStatus, FileMode, CMode (..), UserID, deviceID, fileID, fileMode, fileOwner, getFileStatus, getRealUserID)
 import System.Posix.Files
 import System.Process (CreateProcess, cwd, proc, waitForProcess, delegate_ctlc)
+import System.Which (staticWhich)
 
 import GitHub.Data.GitData (Branch)
 import GitHub.Data.Name (Name)
@@ -241,16 +244,23 @@ inImpureProjectShell shellName command = withProjectRoot "." $ \root ->
 
 projectShell :: MonadObelisk m => FilePath -> Bool -> String -> Maybe String -> m ()
 projectShell root isPure shellName command = do
+  let nixPath = $(staticWhich "nix")
+  nixpkgsPath <- fmap T.strip $ readProcessAndLogStderr Debug $ proc nixPath ["eval", "(import .obelisk/impl {}).nixpkgs.path"]
+  nixRemote <- liftIO $ lookupEnv "NIX_REMOTE"
   (_, _, _, ph) <- createProcess_ "runNixShellAttr" $ setCtlc $ setCwd (Just root) $ proc "nix-shell" $
      [ "default.nix"] <>
-     -- Keep $NIX_PATH in the env for --pure shells so '<nixpkgs>' works in sub-commands
-     -- TODO: Don't use <nixpkgs> for anything!
-     (if isPure then [ "--pure", "--keep", "NIX_PATH" ] else []) <>
+     ["--pure" | isPure] <>
      [ "-A"
      , "shells." <> shellName
      ] <> case command of
        Nothing -> []
-       Just c -> ["--run", c]
+       Just c ->
+        -- TODO: Escape nixpkgsPath and nixRemote
+        [ "--run"
+        , "export NIX_PATH=nixpkgs=" <> T.unpack nixpkgsPath <> "; " <>
+          maybe "" (\v -> "export NIX_REMOTE=" <> v <> "; ") nixRemote <>
+          c
+        ]
   void $ liftIO $ waitForProcess ph
 
 setCtlc :: CreateProcess -> CreateProcess